refactor(memberEntities): moved member related entities to common folder
