+++ src/main/java/pl/uj/passgo/models/member/MemberType.java
@@ -1,4 +1,4 @@
-package pl.uj.passgo.models;
+package pl.uj.passgo.models.member;
 
 public enum MemberType {
     ADMIN, USER, ORGANIZER
